Method fixed

+++ src/main/java/com/handtruth/lessons/lesson3/task1/ForWhile.java
@@ -7,7 +7,7 @@ class ForWhile {
      * <p>
      * Вычислить факториал
      */
-    int factorial(int n) {
+    long factorial(int n) {
         // TODO your code
         return 0;
     }
